✨ chore: split files into route and utils
